chore: refactor modal components
